Spruce up blog overview
This fixes the order of blog posts on the blog overview page so that it
goes from newest to oldest. It also adds the image to the display.

To simplify code we re-use the RecentBlogPosts component in a slightly
altered form. It's probably a good idea to split this out in the future
in some way but still re-use the query and possibly pagination logic.