Change teaser descriptions
Instead of using an excerpt from the body we use the description that's
also been written for search results and social media. This should
convert better.

The date is also moved above the title to better dinstinguish it from
the new teaser description. The elipsis is no longer present to make
this distinction obvious.