Add statistic reloading and reverse sorting
The statistics can now be reloaded without refreshing the page, which
prevents the credentials from having to be updated. This will hang on
the loading state if there's an error though, that should be fixed in
the future.

Additionally clicking a column multiple times will now toggle the sort
order of the column.