Prevent buttons jumping when slides are loading
Unfortunately we don't know the PDF height ahead of time for all of our
presentations. This means we can't set a fixed height which causes our
presentation container to resize when a slide is loading (because
there's no content to render). This causes the buttons to move around
and makes moving through slides quickly difficult.

To combat this we wait for the first slide to be loaded and then set the
height of the presentation container to be that height. This ensures
that even if a slide is loaded and the contents becomes smaller the
buttons will not jump around so users can easily navigate a lot of
slides at once.

We work with the assumption that the size of slides are consistent
throughout a presentation.