Remove yarn as package manager
We're switching back to NPM :) There's no real good reason for this
except that NPM has become a lot better over the years and Yarn seems to
want to complicate things a lot without real benefits. The original
authors for Yarn and the reason I migrated from npm to Yarn also haven't
been with the project for a few major versions.