Ensure larger file uploads work properly
We configure PHP to allow us to upload larger files. We expect that
we're the only ones to actually do this and we want to be able to not
think too much about the filesize of the things we upload and let Drupal
take care of optimising it for any viewers.