Add initial custom Drupal theme
This adds a custom theme that will be used to make my website look
mostly the same as it already does. We create an initial dirty version
of our header component and change some of the default block
configuration.

We also enable the Menu UI so that we can edit the main navigation and
add a "Home" link to it. We don't exclude this module because we expect
to be able to control the menus in production.