Move from yarn to pnpm
Yarn has all this stuff in the repository and for some reason if (with a
lot of effort) I get `yarn set version` to select the version I think I
used in 2023, then it still tries to update the lockfile and is unable
to install packages without making changes. That means that yarn is
unreliable (and I honestly don't understand their versioning).

All the people I see in the front-end community have been happy with
pnpm for years so I make the switch. This has the benefit of removing a
lot of cruft from the repository and getting everything back into
node_modules.

This was done with the latest pnpm version which is 10.5.