Configure CKEditor as editor for posts and talks
Initially I wanted to move to a Markdown editor because it's what I know
how to type. However, Markdown support within Drupal is severly lacking,
so for now we just go with what Drupal provides and then type our text
in the site.

This is a bit more work to migrate existing content but a lot easier in
development time to get the migration to Drupal going.