Provide tools to create homepage
This adds a new page node type using layout builder that can be used to
create the front-page as well as pages such as 404 or my resume. To make
things more visually appealing we add a quick and dirty card
implementation for our post teasers which can be adjusted for talks as
well.