Lint and fix pcss files in theme
This ensures the lint command in the theme focusses on the source files
and not the build output. It also fixes all the issues that do not
change the names of the classes. The expectation is that this does not
change how the theme actually looks but improves compatibility in
browsers.