Fix inconsistency in card styling
There were inconsistencies in the margins of the cards. The margins on
mobile are now missing altogether, but at least they're consistently
missing everywhere, rather than the first card being wider than the
subsequent cards.

A todo for the future is to make sure there are proper margins
throughout the design on mobile.