Ensure site name is in RSS feed names
When importing the RSS feed to external services it would only say
'Posts' or 'Talks' and not have any context that it was from my site. We
fix this by changing the title for that view.

The main view's title does not need to be updated because the Metatag
module adds our site name in the HTML title.