Fix image display for posts full display
Images are now displayed as wide by default when they're rendered. For
our post teasers we're already rendering them as a thumbnail which
reduces the image size.

We also change from "lazy" to "eager" loading by default because if
images are not displayed as thumbnails then they're likely displayed
above the fold.